Try out diff evaluation method

+++ app/src/main/cpp/AudioEngine.h
@@ -92,7 +92,7 @@ public:
     void regenerateTarget();
     std::string getTargetValues();
     void initializeUserPatch();
-    float evaluatePatch();
+    int evaluatePatch();
     void reset();
 
     OutputTerminal outputTerminal;
